***Some Significant Changes relating to an Island Hopping mechanic, Commonwealth capitulation and liberation mechanic, and a DDay mechanic***
Still needs some checking and probably balancing, but putting up at least an initial version for now. Will explain further below on the gitlog. The intention is to probably eventually add some other things like these for Sealion for example.

The intent of these is mainly to help ensure that the Allies have an adequate incentive to fight Japan given that Japan will be expected to normally take the Dutch East Indies and Phillipines and the Allies will normally build refiniries for synthetic rubber, and that the allies have less of an incentive to simply try throwing everything into an early ill-prepared DDay instead. This probably won't be ideally achieved at first, but will be adjusted after experience in games.