The purge is a bit harsher, in particular when Axis has not yet crossed the Stalin line, and in particular with regards to air (also in particular in particular with regards to air in front of the Stalin line). Most/all of this goes away with geography as the Soviets get pushed back, but the hope if it is WAD is that in combination with the Poland infra buffs it should be easier for Axis to push up to and through the Stalin line, but still worthwhile for Soviets to put some troops there to at least delay and damage infra.
However, to help make sure that the late game qualitative advantage does not shift too much to Germany, the Production Targets penalty to switching to more modern equipment is removed (for example so Soviets can more easily switch from 1941 Medium tank to 1943 Medium tank). The hope is that this will partially address some concerns that Soviet modifiers/numbers don't match up well enough.

It remains relatively less efficient for Soviets to switch production because of the higher efficiency cap/low growth rate, but is less so.