France seems perhaps a bit too strong in its ability to hold for long against Germany. In games played with the new France changes where France was actually trying to defend itself, France held for more than 10 weeks twice and for more than 7 weeks once. Given that France is now trying to seriously focus on just defending mainland France and that France currently starts with 24 mils, "hold France" builds are probably a bit more viable than they really should be. So...
France -4 starting mils and -1 starting civ
Germany +4 starting mils and +1 starting civ

The civ is moved in addition to the mils so that it more or less offsets the effect on consumer goods for both countries.

Soviets +2 starting mils to help more or less offset the higher German starting production (also means Soviets can make probably make just a bit more lower-quality divisions with 1936 equipment and be able to defend in depth just a little bit better).