***The Zharak Memorial France Update***
Every state in mainland France starts with level 4 infrastructure, for the purpose of discouraging players of building their factories in mainland France for the purpose of trying for an all-in strategy of holding France outright. Since it is as fast to build in the colonies as in mainland France (more so once you finish the infra focus and get some level 6 infra), you may as well build in the colonies.

When Germany does the Fate of Czech focus, however, infrastructure in France gets restored to vanilla levels automatically via a hidden event, so that supply will be normal for the Battle of France.