***Sealion Naval Defense Mechanic (for testing only so far)***
This won't be actually implemented immediately in the main version of GDU until people have had a chance to take a look at it and make sure it is working properly.

However...

A new mechanic is added which eliminates UK's static 30% core attack/defense bonus and replaces it with a dynamic core attack/defense bonus that depends on the naval situation. Specifically, it depends on how many (of what types) of ships the UK has in the English Channel at the time that a Sealion is occurring.

In the top bar, just to the right of the air experience box, there is a box that shows your current Sealion Naval Defense Score (capped at 70). When UK moves some ships in and out of the English Channel, the Sealion Naval Defense Score goes up when you move more ships in, and goes down when you move them out of the English Channel. If you look at the UK national spirits, you can see that when UK has ships in the English Channel, UK get a core defense and attack bonus for its troops that is equal to your Sealion Naval Defense Score.

For the moment this bonus is set so that it should be pretty strong and goes up to 70% core attack/defense fairly easily (wanting to err on the side of Sealion being too hard rather than being too easy). The entire UK starting fleet is more than sufficient to get the max bonus if it is all in the English channel.