Fuel silo capacity is reverted fully to vanilla (100k per silo). Tornado used to have them set at 50k with the rationale that this was needed for Axis to care about Africa, and then I increased it halfway back up to 75k. That didn't seem to have any impact on Africa being important for Axis, so it is increased again back to vanilla levels.
The Italian silo decision has some silos taken out to more or less make up the difference, so it gives about the same fuel capacity as before. This does mean, however, that if Axis is losing Africa, it is more viable for them to build silos to help stock up more. It also means Japan can stock up on fuel a bit more easily.