The custom graphics and names for the technologies for infantry equipment, AT, ART, and AA are removed. The reason for doing this is gameplay. Although the custom graphics and equipment names are nice, they unfortunately make it impossible to tell which equipment is 1936 inf equip and which is 1942 inf equip (whereas the vanilla graphics do let you distinguish that, i.e. all the vanilla 1918 inf equip graphics have a bayonette and all the 1942 vanilla inf equip has a bazooka, so you can tell which is which).
By the time you have been at war for a while and have scavanged equipment from 20 different countries, you can't tell what's what any more. That means it is not possible to, for example, set your tank template to only use 1942 inf equip. It also means it is not possible to selectively LL out-of-date inf equip (or alternatively to only LL up-to-date versions, for example if you want to LL good equipment to another country for use in their tanks).

If you don't care about those gameplay drawbacks and want to continue using the flavor graphics despite the difficulty of telling which equipment is which, you still can use the same graphics. Just download the graphics mods here.

https://steamcommunity.com/sharedfiles/filedetails/?id=924382320

https://steamcommunity.com/sharedfiles/filedetails/?id=885410896

Since they are graphics mods, they are multiplayer compatible, so you can still use them with GDU if you prefer having them; but if like me and various other people you don't like the gameplay disadvantages, you are no longer forced to use them.

This change does NOT affect equipment types for which these gameplay issues are not really a problem in the same way, such as ships, mot/mech/etc. Just inf equip, art, at, and aa.