Some previous changes to supply/infra in Iraq and the Levant that were made because Tornado wanted it to be easier for Axis to attack into Iraq and then Iran are reverted. The logic was that Axis really needed to push all the way into Iraq and Iran to get all of the oil; however this doesn't seem to be really needed. Axis seems to pretty much have enough fuel even without Iraq. With it reverted it should be substantially less the case that the front automatically goes to Iran as soon as Suez falls (assuming it does). It doesn't really seem to be necessary that it go all the way to Iran every game; the Iran front can be dialed down a bit while remaining possible.
Axis can still get back most of the supply that it had previously in Iraq/Iran (except for the locals supply from VPs in Iraq, which is reverted) and basically push as far into Iraq/Iran as before, but it just has to actually build the infrastructure in order to get most of that extra supply. In particular, Axis can get better supply for attacking into Iraq by building up the infra in Lebanon/Syria, if Axis wants to do that. But Axis does now actually have to pay the price to do that... to pay the price of the Iran front, if you will.

If Axis is not willing to pay that price, then it will have quite bad supply in Iraq and Iran. To some extent this hurts the Allied supply as well, but not as much. Iraq should be substantially more defensible for the Allies than before, because Allies will have a substantial supply advantage there unless/until Axis pays the price to build up the infra.

The airfields do, however, remain Axis favored, so it is not too easy for Allies to push Axis out of the Levant later on. Still, relatively more of the fighting should end up taking place in Levant/Egypt than in Iraq/Iran.