Since vanilla also did something sensible to field marshall/general attack skill values (2.5% per level, down from 5% in vanilla), reverting to vanilla on that as well. GDU had those at 3%, but vanilla's new 2.5% is quite close.
Nevertheless, that means a bit smaller modifiers, and in combination with the vanilla aggressive assaulter change in general this likely makes it a bit harder to attack, so inf equipment and mechanized defense are reduced very slightly to compensate.

Taken together, this shouldn't have a discernable impact, but brings GDU closer to the new vanilla 1.7.1 changes.