1) Navy admirals no longer gain XP levels from battles at all. They just stay the same starting skill level for the whole game, same as with admirals, to prevent admiral farming.
2) The benefit of veterancy to ships is substantially nerfed.

In combination, this should help to stop the naval war from snowballing a bit less; before, if i.e. the UK fought and won a battle against Italy, then it later could use its fleet against Japan and its whole fleet would be veteran with huge bonuses and it would have better admirals, which meant that Japan could not compete in that case. Basically, winning an early naval battle shouldn't make your fleet OP for the rest of the game to the same extent as before.