Reluctantly, the bonus to org from mobile warfare left side tech Mass Motorization is changed to apply only to mot/mech, rather than also to leg infantry.
This was previously the case in GDU prior to the MW buff (with the combat width change), but I didn't realize it had been changed from vanilla. This seems to have been a substantial part in the revival of 20w pure inf, which is not really supposed to be a viable thing in GDU.

While I would ideally like to keep the doctrine changes relative to vanilla as simple as possible so it is easier for newer players (other than DB), it seems like reverting this to how it was before the MW buff is needed. This was an inadvertant change in the first place, and MW is supposed to be good for tanks - not good for pure infantry.