Some defense is shifted from infantry equipment to ART/AT/AA.
The effect of this should be that pure infantry has somewhat lower defense, but quality infantry has similar defense as before, depending on the template.

The shifts are done according to a formula that is hard to explain clearly (I could if I tried, but too much effort at the moment), but is designed to retain the previous balance between 1936/39/42 techs.

Tank/SPG/TD/SPAA defense is also increased because lower inf defense lowers motorized defense, to try to keep tank division defense about the same as before.

There is also a modest increase in infantry breakthrough, but it is still very low relative to other things; so infantry may be a bit more able to attack, but not very much.