According to Florri's suggestion, Greece and Yugoslavia are automatically split up between the Axis along historical borders (with Italy getting Croatia so as to not create an AI and so as to annoy Zharak). The overall goal of this is to try to make sure that minors get some build slots (and cores in the case of Bulgaria) but Germany gets industry that is needed partly for air since normally only Germany and Italy build planes. This also reduces variation in balance between games depending on whether Germany demands to be given everything, or allows minors to have more territory; it will always be the same from game to game now instead.
Most of the starting civs/mils are moved into the areas that Germany will get, on the logic that Germany needs most of the industry to build planes/tanks.

Minors get build slots, and the small amounts of steel are moved into the areas that they will get.

Aluminum is split between mostly Italy and a also some Germany, on the assumption that both need it to produce planes.

Dockyards are moved into the areas that Italy will get.