1) Japan doesn't get 5% extra resources that I added; it seems like they have plenty. Rubber is moved so they still have 16 to export at the start for German starting planes.
2) Japan gets -5% civ construction speed and +5% dock/mil construction speed. The idea is to try to make Japan strong early game, but a bit weaker late game. It seems like they may be too strong in '43 or so.

3) USA gets another 5% civ construction speed from Neutrality, and gets 10% efficiency growth and 10% dockyard capacity. This is to help ensure that USA is sufficiently strong late-game in '43 or so.

Some additional starting factories (particularly mils) for the Balkan Axis minors. They start with less mils in particular than vanilla, which is a strong contrast to the Allied minors. The hope is that this makes them more fun to play and also helps ensure that Axis has enough infantry to garrison stuff etc/have some more reserves/defense in depth in Russia.

4) Bulgaria +2 starting mils and +1 starting civ

5) Hungary +2 starting mils

6) Romania +2 starting mils and +1 starting civ