Finally doing something with some naval changes from about a month ago from before the 1.7 patch that got delayed by 1.7...
The effect of these are:

a) Ship resource costs depend only on the hull, not on the module. The purpose of this is to prevent extreme ships that cost i.e. 1 steel or alternatively 9 steel, and to regularize things so that you don't have strange resource incentives to only use certain types of modules such as secondary batteries or whatever else. An exception to this is heavy cruisers, whose guns cost 0.5 steel, because that is the only way to differentiate light cruisers from heavy cruisers, since they share the same hull. The hope is also that incentives are right in making it worthwhile to research and produce new ship types, rather than produce old ones for perverse resource reasons etc.

b) Ship speeds are more regularized and should not go to such great extremes. In part this is because modules like guns have much smaller effects on speed (which is a reason why certain tech such as the advanced heavy guns are normally never researched at all in vanilla).

The overall intention of this is to make it potentially actually worth researching all (or most) of the naval techs, and more possible to research more of them; hopefully also including things like armor that people don't normally bother with, although more balancing is probably needed.

Kommandanter/Watt/etc may make some further adjustments to this as needed.