Some navy defines changes going along with the previous commit.
Essentially, gun hit profiles are raised by a large amount which has the effect of more or less reducing hit probabilities by a factor of about 10, but the define COMBAT_BASE_HIT_CHANCE is raised by a factor of 10. The end result of that is that overall damage should be more or less the same.

Why do that, you ask, if it makes overall damage about the same? Because in vanilla there is a lower bound of speed where speed no longer makes any difference at all to hit chance. With this change, hit probabilities should depend on speed more continuously now, without hitting a firm lower bound at a certain low (but not that low) speed. Thus, there is an advantage not only to having i.e. 50 speed as compared to 45 speed in terms of your chance to dodge light attack, but there is also an advantage to having 20 speed as comapred to 18 speed (which is not the case in vanilla).

The gun hit profiles also *should* make things more balanced between heavy attack, light attack, and torpedoes (although this was last tested by Kommandanter a while ago prior to 1.7), so that it is not always the best thing to just do pure light attack spam. When this was tested a while ago, you sometimes seemed to get better results with more light attack and other times better results with more heavy attack/torpedoes, depending on what you were fighting. Also i.e. heavy cruisers were sometimes better than BBs, and sometimes BBs better than heavy cruisers, again depending on what you were fighting and what sorts of screens you had.

MIN_SHIP_COUNT_FOR_TASK_FORCE_ROLE_ASSIGNMENT and ADMIRAL_TASKFORCE_CAP are QOL changes from Kommandanter.