USA gets 40% core attack/defense from the Home of the Free national spirit.
If this is not enough to be able to thwart an invasion of the USA, then probably the game is in a situation such that the USA sort of deserves to be invaded; i.e. they probably have no navy and/or army to speak of, and are not training divisions that could be deployed to stop the invasion.

USA's cores on Hawaii etc are ignored after the Pearl Harbor/Tora Tora Tora spirits run out, so that USA will not get these bonuses for defending places like Hawaii, and Midway, but only for the mainland USA.