Soviet industry gets automatically moved to the Urals beginning at the start of Barbarossa.
This works similar to the France "missions" where Germany and the Soviets compete over civs/mils while Germany tries to kill France fast, i.e. using the "missions" system.

1 mil or 1 civ is re-located to the urals/far-east/caucusus each time the mission countdown clock is completed. Mils are evacuated first from each state, then civs.

The # of days it takes to move each factory depends more or less on the normal amount of factories that there should be in each state and how close (or far) each state is from the border.

For example, in Brest (where there is a level 5 fort now from the Stalin line focus), since this is right on the border and since there will be 4 civs there assuming that the player builds none there, it takes 2 days to move each factory (1 for each mission, and 1 day for the mission to update). So basically, in order to evacuate all 4 of the civs, the Soviets need to hold out in Brest for 8 days... And if they only hold out for 2 days, then only 1 of the 4 civs is evacuated.

Likewise for places a slightly further away, (for example Minsk), there should be 7 factories there, but since it is further away it is set so that a factory gets evacuated every 5 days. So the Soviets have to hold Minsk for 35 days to evacuate all 7 factories from Minsk.

So Soviets are incentivized to try to hold all the places where they have factories for as long as possible - or at least for long enough to evacuate most or all of the factories.

All the Soviet focus civs are set to spawn in front of the Stalin line, in Kiev/Minsk/Dnipropetrovsk/Odessa, and a small # of Soviet factories that previously started the game in the urals (not many, just about 2 to 4 or so of them) are moved forward as compared to their  a bit (some others are moved around from the Baltic states to Brest-Litovsk in particular).

When factories are relocated, they are partially "damaged" and need to be "repaired" for a bit before they can be used again. Soviets get a bonus to factory repair speed from the Move Industry to the Urals focus.