There have been some complaints that infantry (infantry equipment is too expensive). So, infantry equipment cost is lowered back to what it was after it was raised previously the last time.
However, if infantry equipment were cheaper without changing the cost of anything else, this would be a buff to pure-inf org spam by making pure infantry in particular cheaper, more so than anything else. So that it is a buff to infantry generally rather than especially to ***pure*** infantry, the cost of AA/AT/ART/support equipment are also lowered again by the same amounts as before.

However, if infantry in general were cheaper but the cost of tanks were not changed, this would lead to a general proliferation of more ground units. To avoid that, tank/armor cost is increased to compensate.

Generally the effect of this should be to make infantry relatively more cost-effective in comparison to tanks than before. If you are having difficulties producing enough infantry equipment or enough infantry, it should be easier to have enough to cover the fronts you need to cover.