Some org is redistributed from infantry to other things - principally to support companies and line AA, but also to other things like ART/AT to a lesser degree.
Specificis:

1) Infantry base org reduced 15% from 55% to 40%. This can be increased further from doctrine, of course.

2) All support companies get an extra 25% org.

3) Line AA gets an extra 40% org.

4) Artillery and AT get an extra 5% org.

5) Mot/Mech base org reduced by 24% (this is nominally more than the increase in infantry org, but this higher increase is necessary to keep org for normal tank templates basically unchanged, because the extra org from support companies etc has a greater effect on tanks.

The overall effect of this is that pure infantry should have significantly less org (15 less org, in fact).

However, the org of quality infantry and tanks should be basically unaffected, quality inf org should be probably slightly higher than before, depending on the exact template.

The goal is that this should discourage pure infantry spam, but not in a way that really encourages excessive/autistic stacking of art/at.

If after this people still think that pure infantry spam is the way to go rather than combined arms warfare with quality inf combined with tanks and air, maybe we can try Baddo's idea of increasing base combat width to 120 or something like that.