Maximum Sealion naval defense bonus of core attack/defense to UK troops raised fom 70 to 100. Each ship counts the same as before, but you don't get capped out as easily.
The UK fleet at the start of the game is large enough to count for 94 sealion naval defense score.

If you have a large fleet near the UK and don't go wandering off with your home fleet over to Gibraltar etc, then it really ought to be impossible to be sealioned with that, at least you would think (though maybe UK infantry still dies to German tanks even with 100% core attack/defense bonus?).