1) A side effect of making mech not require rubber is that Axis needs to make fewer refineries. In the last game, Axis was using about 20% of the rubber they were using for mech, and 80% for planes (Axis was also wasting some rubber). So refinery cost is increased from 16k to 20k.
The effect of this should be that it is similarly costly for Axis to build refineries for planes, given that they don't need rubber for mech any more. Fewer refineries will be needed since you don't need rubber for mech, but offsetting that, each refinery costs more. Another side effect of this is that it is more costly for the Allies to simply ignore Japan and build refineries, which doesn't seem like a bad thing. It also means that it is more costly for Axis to produce fuel from refineries, which also doesn't seem like a bad thing - it incentivizes fighting in Africa, while meanwhile the silos have seemed to keep Axis fuel sufficiently ok enough in recent games after Torch.

2) Soviets start off with 8 less rubber since they don't need it for MOT at the very start of the game. But they get it back from their rubber focus pretty quickly.

3) Soviet supply consumption reduction buffed slightly from -10% supply consumption to -12.5% consumption, because Soviets seem to sometimes be running into worse supply issues than Axis at a certain point, in particular when Axis nears places like Stalingrad/Gorki/etc which starts damaging the high infra supply zones Soviets need for supply.