1) Some issues with Spain fixed:
a) A bit more manpower (will adjust up or down after the first game depending on if it is too much or not enough for reasonable use). Spain gets up to 900k manpower, and more if they are willing to change their draft law to suffer economic penalties. If you think this is not enough, first try not grinding so hard in the Spanish Civil war and let German/Italian vols do more of the attacking.
b) Spain's refineries and oil don't spawn in the impassable Balearic Islands etc.

2) USA's ships that get destroyed by Pearl Harbor are changed to be specifically BB1s, cruiser 1s, and DD1s, rather than BBs, cruisers, and DDs in general, so USA can deploy more modern BBs etc without fear that they will be the ones that get destroyed in Pearl Harbor.