Spain can't make infantry or cavalry. This allows Spain's manpower debuff to be removed without risk that Spain will spam a bazillion infantry. Spain's divisions will be limited simply by special forces cap (which seems fine as it is; you can have about 12 normal special forces divisions, or more if people use templates with more art/at/aa, which they presumably will).
Spain can build as many tank/mot/mech divisions as it wants, limited only by its production.

If Spain wants to make pure AA divisions they can, but this will not be an efficient use of resources for the Axis as compared to LLing excess equipment to Germany or simply making more tanks, and should contribute to Axis losing the game, so probably Germany will not be very happy if Spain does that.