Deep Battle Tweaks -
A little bit of org and HP are redistributed from ART/AT/AA/tanks to inf and mot/mech. The purpose is to ensure that people are not overly incentivized to have very small #s of infantry battalions and mot/mech battalions for tanks (the HP/org were put to the current extreme in an effort to discourage pure inf and encourage quality inf, however that purpose seems to have been largely achieved and perhaps overshot a bit to the extent of encouraging overstacking of things like art/at/aa/tanks/katyushas).

Also, the base research cost of 1943 AA/ART/AT/RART are increased by 50%, to make it a bit harder to tech rush the very last techs.