Trying to get to the right navy balance, based on some input (some of it conflicting) from different people about what is OP/too easily spammable.
1) SHBB modules changed so that you can't put too much AA/secondary batteries on it (which may defeat the intent of the visibility increase in making SHBBs vulnerable to air).

2) The nerf to BCs is modified slightly from -40% heavy gun attack to instead be -35% heavy gun attack and -15% piercing. The idea is to make BCs still be good specifically at killing heavy cruisers, even though they are nerfed against i.e. Battleships/SHBBs.

3) Heavy gun hit profile increased a bit more from 325 to 340, to try to avoid this meaning that everyone simply spams BBs.

The end goal is to promote some sort of balance in what ships are built, but this requires some working out. Although there are unfortunately a lot of frequent changes at the moment to which ships are nerfed/buffed, we are moving quickly in making changes to try to achieve balance more quickly, trying to get ahead of things and pre-emptively nerf some things that are likely to be too strong. So this is why for example SHBBs are being nerfed right after BCs were nerfed, to try to pre-empt a SHBB meta. And this is also why BBs are also nerfed indirectly via the general heavy gun profile increase; to try to pre-empt a regular BB meta.

Hopefully we arrive at a *balanced* navy meta more quickly this way and can get through this period of somewhat tumultuous navy change.