1) The commonwealth's air idea (i.e. Canada's "The Plan") is changed to provide the same bonuses that UK gets from Finest Hour. The same is done for the similar/identical national spirits for Australia and South Africa.
2) The UK's Finest Hour national spirit no longer runs out, it is permanent.

The idea here is for Canada air controller to not provide better bonuses than UK air controller. Instead, they provide the same bonuses. Hopefully this means that Canada air controller can be an option, but that Canada players are not forced into being air controllers. Remember, minors are supposed to be fun to play.

Even without this extra special advantage over the UK, there remain some quite significant advantages of Canada air controller. For example, resource efficiency (Canada has lots of aluminum but not lots of steel/tungsten for tanks), and also convoy efficiency (it takes 0 convoys for USA to LL planes to Canada).