***Norway: The Marsa Matruh of the North***
If Papua New Guinea is the Marsa Matruh of the East, what is the Marsa Matruh of the North?

This is an attempt to make Norway a more active front (and also, within limits, more "mobile", at least in the sense of more push-able).

The area around Narvik is made 2-tiles wide rather than 1, and there is a continuous path to Narvik of hill tiles (also with the other tiles being mountains. If Allies go into Norway/Narvik in 1939/40, this should make it harder for them to permanantly hold it without eventually getting kicked out. There is also a continuous hill-path to Bergen.

However, infra is also increased slightly, which can allow the Allies to supply a bit more troops there, and also may make it more viable for the Allies to later naval invade Norway in '42/'43/etc. Later on, the fact that there are hill-paths rather than just purely mountains may make it easier for Allies to take back all of Norway if they at first take back part of it.

Minors New Zealand, Canada, Hungary, and Spain also get some changes to generals to make them a bit better suited to fighting in Norway. i.e. Mountaineer infantry generals and Panzer Leader hill generals. This is so that if the Norway front is active, you can potentially give control of the front to one of those minors, while other players focus their attention on other fronts.