Reverting/reducing some of the previous buffs to Deep Battle, mainly less/no additional org from things like ART/AT/tanks, and a bit fewer hit points.
This is a fairly small nerf because I am afraid of over-nerfing and then people doing pure inf again. There seems to be a narrow range on either side of which people start to do either pure infantry org spam or else autistic 10 AT divisions; ideally we want to be promoting neither of those two extremes, but rather encouraging normal quality infantry.