***The IGB memorial France update*** -
*the latest in an on-going series cracking down on French cheese one exploit at a time*

If France doesn't defend the Maginot line and Germany takes it, that doesn't trigger the countdown for the 10 civ/mil factories. So France can't not defend the Maginot line in order to bait Germany to take it to trigger the countdown early. This also applies to Franche Comte, the state right behind the Maginot line.