Criteria for Axis annexing Vichy if Italy is being invaded changed from rather than being just 1% surrender progress of Italy triggering it regardless of date to 1% surrender progress of Italy only triggering it if it is ALSO after June '41, or else 25% surrender progress triggering it if it is also after January '41.
The point of this is not/was not intended to mean that Italy shouldn't worry about defending Corsica/Siciliy, but rather was really just to try to ensure there is not a disincentive for the Allies to do Torch and to just try to invade those directly first instead.

But this way it should still be sensible for the Allies to try to put pressure on Italy by invading stuff prior to Torch (i.e. in 1939/1940 before the USA joins) if they can.