Since Allies recently have often been building large amounts of tanks (up to 16 or so), a potential problem is that the Allies might be able to lend lease all those tanks to the UK, and then UK goes and defends France with that many tanks. This could potentially ruin a game or games.
To prevent that, it is no longer possible for UK to receive LL until either Paris or another French state similarly far-in to France has fallen, or alternatively until either London, Gibraltar, Alexandria, or Cairo has fallen.

In practice this should generally mean that UK can only receive LL after Paris has fallen.

The commonwealth can still receive LL from UK and USA as normal, however.

Effectively this limits any BEF that UK sends to France to being equipped with UK's own equipment, rather than allowing the entire commonwealth + USA to load up UK and pour everything into France. The idea is for a normal BEF to be encouraged, but not some crazy 16 tank BEF to be encouraged (or possible).