Axis countries can license spy on Soviets again.
I figured out that with some altered code I could make it so that only Allied countries were blocked from licensing from Soviets (to avoid civ boosting/rule breaking/licensing from the Soviet designers), while Axis could still view Soviet licenses (to see what they are building).

IMO it is sort of better, or at least arguably so, if NOBODY can see the other side's licenses (so that you can get some additional element of surprise with your build), but as long as some countries can use license spying on the enemy, it makes sense for all sides to be able to do so. The reason for Soviet licenses being blocked was not to block license spying, but to block civ boosting and stopping the Allies from getting the piercing of the Soviet tank designer.