In addition, the buff that Soviets get when the Heartland is in danger (i.e. when Axis has pushed as far as getting to i.e. the approaches to Moscow) is also changed to be based purely on time rather than geography, and is re-named to "Late Barbarossa Balance Adjustment". It happens 85 days after the start of Barbarossa, which should be the same time as Soviets get Defense of Moscow. IMO it should probably be a bit later than that if Soviets defend properly, something like 100-110 days, but for the moment I would rather err on the side of Soviets being harder to kill rather than being too easy to kill, since if they die too easily the game ends early for everyone.
This is done partly because it was always the intent to make these buffs be time-based rather than geography based when the balance became clearer, but also because certain people have been trying to say that all Soviets can do is defend in "their" air zones, and that this is allegedly not fun. Although I don't think that argument is valid (just because it is what Gerardo does does not mean it is all Soviets can do and does not mean it is optimal), this further eliminates any incentive for Soviets to not defend more or less from the front. If all Soviets do is try to defend "their" air zones, they are being pretty dumb, because they are not getting any particular reduction of the purge at that point. Also remember that CAS should fly more or less regardless of what color an air zone is. Air zones don't matter as much as you think they do, and to the extent Soviets were previously better able to defend in their air zone that was partly because this buff was geographical. But now it is not geographical.