Hungary +1 civ
Finland +1 mil
Spain -1 civ and -1 mil

This is to even out the strength of the Axis minors slightly. Currently Spain is by far the strongest since they get so much trade for tungsten. Might move another civ from Spain to Hungary if Hungary is still too weak in the next game.