Medium Tank 1 is a 1938 tech, and Germany gets an ahead-of-time bonus so it can get it without an ahead of time penalty as soon as Germany does the tank treaty focus.
This should cut down somewhat on the possibility of Allies having ridiculous #s of tanks (mostly '37 mediums) too early in Egypt.

If it is needed, some sort of ahead-of-time penalty reduction (half a year?) might be given to Italy and/or UK, but probably not commonwealth, so that UK and Italy can build some early. But let's try it like this for now, it also makes it a bit closer to vanilla, which is a good thing ceteris paribus.