All the changes that RikiBreeiki lists on the 8.8 discord as being "differences from GDU" account are implemented in GDU itself. Riki lists those changes as:
"1- Bulgaria Finland retain their civs, finnish civs only moved if it is ai
2-10 nations have 7 extra focuses, 3 mutually exclusive with other 3, called talent tree
3-Capitals take double damage from torpedos, Torpedos and light guns are slightly more accurate
4-No revan's france changes in 8.8
5-1940 bb and its' needed techs are 1939 techs
6-No Nazi flag"

In the case of the meme talent tree focuses, those can be either enabled or disabled by the host (added as a new option to either enable or disable the focuses). This way people who want to play a serious game can play with them disabled, and people who want to play a meme game and joke around can play with them enabled, according to their choice (currently they appear/disappear and I think it might take a hotjoin to update, this may be changed to be "available"/"unavailable" if this doesn't work right).

As I told Riki (and in fact told him several weeks ago), he could make what changes he wanted to GDU rather than making them in a separate mod. But for whatever reason, he didn't do so, and instead made a parallel mod.

There is really no valid/sensible reason to have duplicate mods that are so close to being the same, but use different gits and have a few differences that are hard to pin down. All this does is make it harder to get balance right, since some games will be played with a slightly different version, and eventually after the mods further diverge it will be impossible to really tell how much of what happens in games is due to slight differences in the mod. It also makes it take a lot more work to update things, because you have to copy things from one mod to the other to keep them both more or less up. This is basically exactly how you should not use git, it is horrible modding practice and wastes a ton of duplicate effort. As such, it is harmful to the mod and also to getting good games for the community of players.

So I ask Riki to please make whatever changes need to be made in the main GDU mod (as he is allowed to do and has been allowed to do). If he nevertheless persists in making a parallel mod that is almost but not quite the same as GDU and steals basically everything that GDU changes, I will take it as a personal insult.

There are some small differences still that Riki didn't think important enough to mention in his list of differences from GDU, if he wants to incorporate these he can go ahead and do so. The only reason I didn't already is that it is not clear exactly what they all are to me (they are split among a bunch of his different commits and various updates - this is precisely the problem with trying to mod with 2 different parallel gits; it just is not functional/sensible in any way, shape, or form).

In sum, there are no real differences between GDU and Riki's parallel version now, and there is no reason to not play with the main version of GDU (and there is no reason for Riki to put his changes into a separate mod, while copying ongoing changes from GDU itself).