When France falls, Germany gets resource rights to about half of the Spanish tungsten (Andalusia state), but 5 civs are transferred from Germany to Spain.
The intent of this mostly to help solve the issue of Spain running out of owned civs and being unable to trade for resources in the late game (because it gets so much trade), and also to make it less game ending if the Spain player is not the best. Spain should still be important though because Germany should run out of that tungsten at a certain point, but should get less trade.